requirements.txt
# Core dependencies for the RAG system using PostgreSQL pgvector
langchain-postgres>=0.0.1
langchain-google-genai>=0.0.5
psycopg[binary]>=3.0.0
pandas>=1.5.0
tqdm>=4.64.0

# For environment variable management
python-dotenv>=0.19.0